# openapi.yaml
openapi: 3.1.0
info:
  title: BeeGFS Cluster Management API
  description: REST API for monitoring and managing a BeeGFS-inspired distributed storage cluster.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Development server
paths:
...
  /jobs:
    get:
      operationId: getJobs
      summary: List jobs
      description: Returns all cluster jobs (queued, running, completed, failed, cancelled).
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      summary: Create a new job
      description: Creates a new cluster job (rebalance, pool expansion, or health check).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  $ref: "#/components/schemas/JobType"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    delete:
      operationId: cancelJob
      summary: Cancel a job
      description: Cancels a queued or running job.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Job cannot be cancelled (already finished)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    JobType:
      type: string
      enum: [rebalance, pool_expansion, health_check]
    JobStatus:
      type: string
      enum: [queued, running, completed, failed, cancelled]
    Job:
      type: object
      required: [id, type, name, status, progress, createdAt, updatedAt]
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/JobType"
        name:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        progress:
          type: number
          minimum: 0
          maximum: 100
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        error:
          type: string
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string